+++ src/node_device/node_device_driver.h
@@ -39,6 +39,7 @@ typedef enum {
      * separation makes our code more readable in terms of knowing when we're
      * starting a defined device and when we're creating a transient one */
     MDEVCTL_CMD_CREATE,
+    MDEVCTL_CMD_MODIFY,
 
     MDEVCTL_CMD_LAST,
 } virMdevctlCommand;
@@ -186,3 +187,13 @@ virCommand*
 nodeDeviceGetMdevctlSetAutostartCommand(virNodeDeviceDef *def,
                                         bool autostart,
                                         char **errmsg);
+
+virCommand*
+nodeDeviceGetMdevctlModifyCommand(virNodeDeviceDef *def,
+                                  bool defined,
+                                  bool live,
+                                  char **errmsg);
+int
+nodeDeviceUpdate(virNodeDevice *dev,
+                 const char *xmlDesc,
+                 unsigned int flags);